+++ jgclark.Reviews/CHANGELOG.md
@@ -5,6 +5,12 @@ See [website README for more details](https://github.com/NotePlan/plugins/tree/m
 - ??? Fixed the race condition on (un)pausing a project
 - ??? ability to pause/unpause a project, by calling new **/pause project toggle** command or adding/removing `#paused` to a project's metadata. When paused this stops the note from being included in reviews, but keeps it visible in the project lists.
 -->
+## [0.10.1] - 2023-05-05
+### Changed
+- the 'Folders to Include' setting now matches anywhere in the folder name (for @dwertheimer)
+### Fixed
+- it could fail when running NP 3.9.0 or earlier
+
 ## [0.10.0] - 2023-05-04
 ### New
 - big speed up possible on large collections by specifying new 'Folders to Include' setting. (addresses [#442](https://github.com/NotePlan/plugins/issues/442) for @dwertheimer)